+++ tests/per-field-types/test-types.ts
@@ -1,0 +1,5 @@
+// Import the output file to ensure that its import statements are valid
+import * as customTypes from "./__TEST_TMP__/perFieldTypes.js";
+import * as perFieldTypes from "./__TEST_TMP__/withCustomTypes.js";
+
+console.log(customTypes, perFieldTypes);